Cherry-pick non-enclave bugprone tidy fixes
Curated non-enclave subset of development commit a78d34d5b7b3871944d06fa67ec42f56c12ee2bd. Dropped SGX/enclave paths and restored std::move(params) in rpc::base dispatch to avoid copying send_params payloads.

+++ c++/rpc/src/stub.cpp
@@ -128,7 +128,7 @@ namespace rpc
         RPC_ASSERT(count != std::numeric_limits<uint64_t>::max());
         RPC_ASSERT(count != 0);
 
-        uint64_t ret = error::OK();
+        int ret = error::OK();
         auto transport = zone_->get_transport(caller_zone_id);
         if (transport)
         {
